07-procesos-en-tiempo-real/pipelines/pyflink-pipeline/pyproject.toml
[project]
name = "pyflink-pipeline"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
]
